+++ Changelog.md
@@ -32,6 +32,8 @@ Compiler Features:
 Bugfixes:
  * SMTChecker: Fix lack of reporting potential violations when using only the CHC engine.
  * SMTChecker: Fix internal error on conversion from string literal to byte.
+ * SMTChecker: Fix internal error when using tuples of rational literals inside the conditional operator.
+ * SMTChecker: Fix internal error when assigning state variable via contract's name.
  * Code generator: Fix missing creation dependency tracking for abstract contracts.